+++ src/preprocess/preprocessor.h
@@ -14,7 +14,6 @@
 #include "backtrack/assertion_stack.h"
 #include "backtrack/pop_callback.h"
 #include "preprocess/assertion_tracker.h"
-#include "preprocess/pass/contradicting_ands.h"
 #include "preprocess/pass/elim_lambda.h"
 #include "preprocess/pass/elim_udiv.h"
 #include "preprocess/pass/embedded_constraints.h"
@@ -119,7 +118,6 @@ class Preprocessor
 
   /** Preprocessing passes */
   pass::PassRewrite d_pass_rewrite;
-  pass::PassContradictingAnds d_pass_contr_ands;
   pass::PassElimLambda d_pass_elim_lambda;
   pass::PassElimUdiv d_pass_elim_bvudiv;
   pass::PassEmbeddedConstraints d_pass_embedded_constraints;